backend/pyproject.toml
[project]
name = "kakao-match-backend"
version = "0.1.0"
description = "Kakao Matchmaking Backend API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.5",
    "apscheduler>=3.11.0",
    "bcrypt>=4.3.0",
    "fastapi>=0.116.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.43",
    "structlog>=25.4.0",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]
